+++ src/m5/M5AtomicMotionExt.cpp
@@ -17,7 +17,7 @@ void M5AtomicMotionExt::outputServoPulse(int channel, uint16_t pulse) {
   setServoPulse(channel, pulse);
 }
 
-void M5AtomicMotionExt::outputMotorSpeed(int channel, double normalized_speed) {
+void M5AtomicMotionExt::outputMotorSpeed(int channel, float normalized_speed) {
 
   if (channel < 0) 
     return;
